Umstellung auf Ogre 1.0 und Cegui 0.2.0 und OgreOde 1.0Rc1 watershed edition
git-svn-id: svn://svn.berlios.de/dsa-hl@642 4c79e8ff-cfd4-0310-af45-a38c79f83013

+++ rl/trunk/engine/dialog/src/Match.cpp
@@ -47,7 +47,7 @@ namespace rl
 	{
 		//--	only need to check index > size (due to unsigned int)
 		vector<string> &vs = stars[which];
-		unsigned int size = vs.size();
+		size_t size = vs.size();
 		//--	++index accounts for default first item (why is it there?)
 		if ( ++index > size ) {
 			//--	throw an exception?
